+++ antifraud/src/main/java/com/bank/antifraud/dto/SuspiciousAccountTransfersDto.java
@@ -1,9 +1,7 @@
 package com.bank.antifraud.dto;
 
-import java.math.BigInteger;
-
 public record SuspiciousAccountTransfersDto
-        (BigInteger id,
+        (Long id,
          Long accountTransferId,
          Boolean isBlocked,
          Boolean isSuspicious,
