+++ antifraud/src/main/java/com/bank/antifraud/repository/SuspiciousCardTransferRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface SuspiciousCardTransferRepository extends JpaRepository<SuspiciousCardTransferEntity, BigInteger> {
+public interface SuspiciousCardTransferRepository extends JpaRepository<SuspiciousCardTransferEntity, Long> {
 }
